[arcane,impl] Déplace l'affichage des statistiques d'exécution des boucles dans 'ExecutionStatsDumper'.
